Problem 101
Problem Title: Turn it to original string

Statement: two friends found unique way to chat that they encode the message and send it to one another.

for the encoding purpose they are using old phone’s keypad in which multiple clicks of same button will generate the different character.

The keypad image is given below.

  1      2         3
        ABC     DEF
  4      5         6
GHI   JKL     MNO
 7        8          9
PQRS  TUV    WXYZ
  *        0          #
            _
keypad has 12 buttons which have digit and characters as well. the encoding technique using this keypad as given below. if we click one time the button printed with 2 consider it as A. If we click two times the button printed with 2 like 22 consider it as B. If we click three-time the button printed with 2 like 222 consider it as C. If we click four-time the button printed with 2 like 2222 consider it as number 2 but there is a flaw with the above technique, that is, if they type 222222 then this sequence can consider as ABC or BAC or CBA or any other combination of A,B, C and 2. In order to resolve this condition they uses a full stop between sequence of 2.

For example 222.2.22 means CAB For the space they are using 0 between the numbers and when they want to write number zero by this encoding technique they uses 00.

For example 20222 means A C and 20.00222 means A 0C. If they click button 1 single time consider it as digit 1.

Your task is to trace the given sequence properly and decode the string in to the character version.

Finally print that decoded actual character string.

Example:

INPUT -> 69990662633044477770442777.777999

OUTPUT -> MY NAME IS HARRY