Problem No. 102
Count apples 

In order to decide the price of apples company uses the size of apple as parameter. Larger the size of apple means price is more.so to pack and filter the different size apples they use a conveyor belt and different size vessels. Suppose all apples moves sequentially over the belt. Vessels with different diameter are attached with the conveyor belt such that when apples move over conveyor belt they fall in vessel if the diameter of vessel is greater or equal to apple diameter.

Suppose there are n vessels attached with belt and m apples are present over the belt. All vessels have different diameter but the same capacity of storing 3 apples. This arrangement has a problem that less diameter size of apples can fall into the greater diameter vessel. Your task is to find due to this wrong arrangement which apples couldn’t fall into the vessels and remains over the belt. To implement the above operation, an array named as V contains the diameter of vessels that are attached to belt in the same order as 
they are available in array. One another array A contains the diameter of apples.

For Example, suppose array V contains elements {7, 5, 3} denoting diameter of each vessel and they are attached with belt in same order as they are available in array. that means vessel V[0]=7 appears first in moving direction of belt. Array A contains elements {7,5,6,6,3,5,3,5,4} denoting diameter of each apple and the same sequence from index 0 to n-1 they will move over the belt. that means A[0]=7 is first placed on belt and A[8]=4 is last placed on belt. So first vessel of diameter 7 filled with apples A[0]=7, A[1]=5and A[2]=6 but now due to vessel V[0] is full. A[3]=6 will not fall in any vessel due to rest vessels have small diameter than A[3]. Second vessel of diameter 5 filled with apples A[4]=3, A[5]=5 and A[6]=3 but now apple A[7] and A[8] can’t fall in any vessel. So the output should look like 3 7 8 denoting positions of apples which are not fall in vessel due to wrong arrangement.

Example:

Input Vessels and Apples -> 3 9

Input Vessels diameters -> 7 5 3

Input Apples diameters -> 7 5 6 6 3 5 5 4

Output -> 3 7 8