Problem No. 122
Title -> Interval Search (Enhanced Linear)



Statement: This search is the enhanced form of linear search. in linear search we search for the element by traversing the sorted

array elements one by one. Which gives O(n) complexity to search the element in worst case.

In this searching technique we also consider a skip or interval value x. so we start searching from first element that is 0th

index element and then we add the skip/interval value in index value to get the next position to search the element.

this algorithm will work on sorted array so, if kth element is greater than the E (element to be searched), reduce the interval by half of its

original value and move backwards to search that element. if interval value is odd that is suppose 3 then consider 3/2=1.

In backward direction if you found (k-x)th element is smaller than E reduce the interval by half again and now search in forward direction.

Repeat the process until you found the desired element.

Remember while moving forward with interval value x you should not cross/exceed the array.

and while decreasing the x by 2, minimum value of x can only be 1.

Finally print the index of element to be searched. if element is not found in array print -1 in output.

Input: first line contains an integers n denoting size of array

Second line contains n integers denoting sorted array.

Third line contains two integers E and x respectively denoting element to be searched and interval/skip value.

Output: contains either index value of element to be searched or -1 if element is not found in array.

Constraints: 10<=n<=1000

2<=x<=10

Input:
10
1 2 4 6 7 9 12 13 25 28
6 2
Expected: 
3

Input:
20
1 2 3 4 6 7 9 12 13 25 23 28 35 43 54 63 76 99 100 110
2 5       
Expected: 
1