Problem No. 120
Title -> Find the correct boxes



Statement: there are n medicine boxes are arranged in sequence in medical store. Each box has a sticker of alphabet denoting

medicines starting from that alphabet is stored inside that box. Also medicines are more so some boxes have repeated alphabet sticker on that.

The only problem is since boxes are placed over some height and by mistake stickers are pasted on top of the box and height of the employee

is not enough to see that stickers. So he uses a stair to read that sticker and find the correct box of medicine.

Due to the above arrangement employee has to shift stairs too many times in order to search the correct box and medicines.

Your task is to help that employee and search the correct box/boxes of medicine in log n stair shift operation.

For example, suppose boxes are arranged in following order that are denoted by string boxes

boxes=aaabcdeeeffghssssttuuwz

and the medicine you have to search is sepia so sepia medicine can be found in box 14 15 16 17.

(you have to print the box positions that is index+1)

If medicine is not found in any box print -1.

Input: first line contains a string of nalphabets denoting boxes with sticker pasted on it.

second line contains a string denoting name of medicine that you have to search in boxes.

Output: output contains positions of boxes separated by space.

Constraint:10<=n<=200

5<=Medicine length<=20

Array/string contains only small case letters.

Example:

aaabbcdeff

ferrom met

9 10